requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic>=2.0.0
python-dotenv>=1.0.0
passlib[bcrypt]>=1.7.4
bcrypt<4.0
pytest>=8.0.0
